+++ src/imio/smartweb/core/contents/rest/events/endpoint.py
@@ -1,61 +1,16 @@
 # -*- coding: utf-8 -*-
 
 from datetime import date
-from datetime import datetime
-from datetime import timedelta
+
 from imio.smartweb.core.config import EVENTS_URL
 from imio.smartweb.core.contents.rest.base import BaseEndpoint
 from imio.smartweb.core.contents.rest.base import BaseService
-from plone.event.recurrence import recurrence_sequence_ical
-from plone.event.utils import pydt
+from imio.smartweb.core.utils import expand_occurences
+from imio.smartweb.core.utils import get_start_date
 from plone.restapi.interfaces import IExpandableElement
-from plone.restapi.serializer.converters import json_compatible
-from pytz import utc
 from zope.component import adapter
 from zope.interface import implementer
 from zope.interface import Interface
-
-import copy
-import dateutil
-
-
-def get_start_date(event):
-    return datetime.fromisoformat(event["start"])
-
-
-def expand_occurences(events):
-    expanded_events = []
-
-    for event in events:
-        if not event["recurrence"]:
-            expanded_events.append(event)
-            continue
-        start_date = dateutil.parser.parse(event["start"])
-        start_date = start_date.astimezone(utc)
-        end_date = dateutil.parser.parse(event["end"])
-        end_date = end_date.astimezone(utc)
-
-        start_dates = recurrence_sequence_ical(
-            start=start_date,
-            recrule=event["recurrence"],
-            from_=datetime.now(),
-        )
-
-        if event["whole_day"] or event["open_end"]:
-            duration = timedelta(hours=23, minutes=59, seconds=59)
-        else:
-            duration = end_date - start_date
-
-        for occurence_start in start_dates:
-            if pydt(start_date.replace(microsecond=0)) == occurence_start:
-                expanded_events.append(event)
-            else:
-                new_event = copy.deepcopy(event)
-                new_event["start"] = json_compatible(occurence_start)
-                new_event["end"] = json_compatible(occurence_start + duration)
-                expanded_events.append(new_event)
-    expanded_events_sorted = sorted(expanded_events, key=get_start_date)
-    return expanded_events_sorted
 
 
 class BaseEventsEndpoint(BaseEndpoint):
@@ -96,8 +51,9 @@ class EventsEndpoint(BaseEventsEndpoint):
             return res
 
         expanded_events = expand_occurences(res["items"])
-        res["items"] = expanded_events
-        res["items_total"] = len(expanded_events)
+        expanded_events_sorted = sorted(expanded_events, key=get_start_date)
+        res["items"] = expanded_events_sorted
+        res["items_total"] = len(expanded_events_sorted)
 
         return res
